+++ controllers/demoRequest.controller.js
@@ -1,7 +1,12 @@
 const { demoRequestService } = require("../services");
+const { validationResult } = require("../validations/demo.validator.js");
 
 const createDemoRequest = async (req, res) => {
   try {
+    const errors = validationResult(req);
+    if(!errors.isEmpty()) {
+      return res.status(400).json({ message:"error to create DemoRequires", errors:errors.array() });
+    }
     const demoRequest = await demoRequestService.createDemoRequest(req.body);
     res
       .status(201)
@@ -14,8 +19,12 @@ const createDemoRequest = async (req, res) => {
 
 const getDemoRequest = async (req, res) => {
   try {
+    const errors = validationResult(req);
+    if(!errors.isEmpty()) {
+      return res.status(400).json({ message:"error to get DemoRequires", errors:errors.array() });
+    }
     const demoRequestId = req.params.id;
-    const demoRequest = await demoRequestService.getDemoRequest(demoRequestId);
+    const demoRequest = await demoRequestService.getDemoRequests(demoRequestId);
 
     if (!demoRequest) {
       return res.status(404).json({ message: "DemoRequest not found" });
@@ -30,6 +39,10 @@ const getDemoRequest = async (req, res) => {
 
 const updateDemoRequest = async (req, res) => {
   try {
+    const errors = validationResult(req);
+    if(!errors.isEmpty()) {
+      return res.status(400).json({ message:"error to update DemoRequires", errors:errors.array() });
+    }
     const demoRequestId = req.params.id;
     const updatedDemoRequest = await demoRequestService.updateDemoRequest(
       demoRequestId,
@@ -54,6 +67,10 @@ const updateDemoRequest = async (req, res) => {
 
 const deleteDemoRequest = async (req, res) => {
   try {
+    const errors = validationResult(req);
+    if(!errors.isEmpty()) {
+      return res.status(400).json({ message:"error to delete DemoRequires", errors:errors.array() });
+    }
     const demoRequestId = req.params.id;
     const deletedDemoRequest = await demoRequestService.deleteDemoRequest(
       demoRequestId
